パン屑現行ページ実装：残CSS調整、SEO
