+++ picker-app/src/pages/track-and-field/index.tsx
@@ -1,6 +1,7 @@
 import { ReactElement } from 'react';
 import { GetStaticProps } from 'next';
 import { shoeEventCategories } from 'picker-types/types/track-and-field/shoeEvents';
+import { TAFBreadcrumbs } from 'features/track-and-field/common/TAFBreadcrumbs';
 import { TAFTopTemplate, TAFTopTemplateProps } from 'features/track-and-field/top/TAFTopTemplate';
 import TAFLayout from 'layout/TrackAndField';
 import { NextPageWithLayout } from 'pages/_app';
@@ -10,6 +11,7 @@ const TAFIndexPage: NextPageWithLayout<TAFTopTemplateProps> = props => {
   return (
     <>
       <TAFTopTemplate {...props}></TAFTopTemplate>
+      <TAFBreadcrumbs links={[]}/>
     </>
   );
 };
