+++ picker-app/src/pages/track-and-field/search/index.tsx
@@ -1,4 +1,5 @@
 import { ReactElement } from 'react';
+import { TAFBreadcrumbs } from 'features/track-and-field/common/TAFBreadcrumbs';
 import { TAFSearchTemplate } from 'features/track-and-field/search/TAFSearchTemplate';
 import TAFLayout from 'layout/TrackAndField';
 import { NextPageWithLayout } from 'pages/_app';
@@ -7,6 +8,7 @@ const TAFSearchPage: NextPageWithLayout = () => {
   return (
     <>
       <TAFSearchTemplate></TAFSearchTemplate>
+      <TAFBreadcrumbs links={[]}/>
     </>
   );
 };
